backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
requests==2.31.0
pymongo==4.6.1
beautifulsoup4==4.12.3
lxml==5.1.0
yt-dlp==2024.3.10
openai-whisper==20231117
newspaper3k==0.2.8
GoogleNews==1.6.14
python-dateutil==2.8.2
